% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_pkgs.R
\name{adjust_py_packages}
\alias{adjust_py_packages}
\title{Adjust Python package names for import statements}
\usage{
adjust_py_packages(packages)
}
\arguments{
\item{packages}{Character vector of package names}
}
\value{
Character vector of adjusted package names
}
\description{
Adjust Python package names for import statements
}
