% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivs.R
\name{build_copy_commands}
\alias{build_copy_commands}
\title{Build copy commands for files}
\usage{
build_copy_commands(files, recursive = TRUE, indent = 6)
}
\arguments{
\item{files}{Character vector of files to copy}

\item{recursive}{Logical, use cp -r if TRUE}

\item{indent}{Number of spaces for indentation}
}
\value{
String with copy commands
}
\description{
Build copy commands for files
}
\keyword{internal}
