% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxp_io.R
\name{build_hybrid_src}
\alias{build_hybrid_src}
\title{Build hybrid source configuration for remote URL + local user functions}
\usage{
build_hybrid_src(url, user_functions)
}
\arguments{
\item{url}{Character URL}

\item{user_functions}{Character vector of additional files}
}
\value{
Character Nix expression
}
\description{
Build hybrid source configuration for remote URL + local user functions
}
\keyword{internal}
