% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxp_io.R
\name{build_local_src}
\alias{build_local_src}
\title{Build local source configuration}
\usage{
build_local_src(path, user_functions)
}
\arguments{
\item{path}{Character path}

\item{user_functions}{Character vector of additional files}
}
\value{
Character Nix expression
}
\description{
Build local source configuration
}
\keyword{internal}
