% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxp_io.R
\name{build_phase}
\alias{build_phase}
\title{Build build_phase command}
\usage{
build_phase(
  lang,
  read_func,
  user_code,
  out_name,
  path,
  user_functions = character(0),
  serialize_str = NULL
)
}
\arguments{
\item{lang}{\code{"R"}, \code{"Py"}, or \code{"Jl"}.}

\item{read_func}{String representing the function to call for reading data.}

\item{user_code}{Source/import statements for user functions.}

\item{out_name}{Name of the output object (RDS/pickle file).}

\item{path}{Input path (file or folder).}

\item{user_functions}{Character vector of user function files.}

\item{serialize_str}{Language-specific serializer identifier or NULL for default.}
}
\value{
A string with the build phase commands.
}
\description{
Constructs the build-phase shell command for R, Python, or Julia.
}
\keyword{internal}
