% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxp_io.R
\name{build_reticulate_phase}
\alias{build_reticulate_phase}
\title{Build reticulate build phase}
\usage{
build_reticulate_phase(r_command)
}
\arguments{
\item{r_command}{Character R command}
}
\value{
Character build phase
}
\description{
Build reticulate build phase
}
\keyword{internal}
