% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxp_io.R
\name{build_src_part}
\alias{build_src_part}
\title{Build Nix src part}
\usage{
build_src_part(path, user_functions = character(0))
}
\arguments{
\item{path}{Character path to file or directory.}

\item{user_functions}{Character vector of additional files to include.}
}
\value{
A string with the \code{src} Nix expression.
}
\description{
Creates the \code{src} attribute for the Nix derivation. Handles both remote URLs
(using \code{nix-prefetch-url}) and local paths (with optional user function
scripts).
}
\keyword{internal}
