% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivs.R
\name{build_src_snippet}
\alias{build_src_snippet}
\title{Build src snippet for Nix derivation}
\usage{
build_src_snippet(fileset_parts)
}
\arguments{
\item{fileset_parts}{Character vector of files to include}
}
\value{
String with src snippet or empty string
}
\description{
Build src snippet for Nix derivation
}
\keyword{internal}
