% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxp_io.R
\name{build_transfer_command}
\alias{build_transfer_command}
\title{Build transfer command for py2r or r2py}
\usage{
build_transfer_command(out_name, expr_str, direction)
}
\arguments{
\item{out_name}{Character output name}

\item{expr_str}{Character expression string}

\item{direction}{Character direction}
}
\value{
Character command
}
\description{
Build transfer command for py2r or r2py
}
\keyword{internal}
