% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxp_io.R
\name{clean_user_functions}
\alias{clean_user_functions}
\title{Clean user functions vector}
\usage{
clean_user_functions(user_functions)
}
\arguments{
\item{user_functions}{Character vector of user-defined function script paths.}
}
\value{
Character vector without empty strings.
}
\description{
Removes empty character entries from the provided \code{user_functions} vector.
}
\keyword{internal}
