% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxp_io.R
\name{create_rxp_derivation}
\alias{create_rxp_derivation}
\title{Create rxp_derivation object}
\usage{
create_rxp_derivation(
  out_name,
  snippet,
  lang,
  user_functions,
  nix_env,
  env_var,
  encoder = NULL
)
}
\arguments{
\item{out_name}{Character output name}

\item{snippet}{Character Nix snippet}

\item{lang}{Character language}

\item{user_functions}{Character vector}

\item{nix_env}{Character nix environment}

\item{env_var}{Named list of environment variables}

\item{encoder}{Character serializer identifier (for display)}
}
\value{
rxp_derivation object
}
\description{
Create rxp_derivation object
}
\keyword{internal}
