% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dag.R
\name{dag_for_ci}
\alias{dag_for_ci}
\title{Export DAG of pipeline and prepare it for rendering on CI}
\usage{
dag_for_ci()
}
\value{
Nothing, writes \code{dag.dot} in \verb{_rixpress/}.
}
\description{
This function is called automatically by rxp_ga(), but
you can also run it yourself if you want to visualize the DAG on CI
with your own workflow definition.
}
\examples{
\dontrun{dag_for_ga()}
}
