% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deriv_read_load.R
\name{drv_load}
\alias{drv_load}
\title{Load a derivation's output into global environment}
\usage{
drv_load(derivation_name, result_path = "_rixpress")
}
\arguments{
\item{derivation_name}{Character, the name of the derivation.}
}
\value{
None. The derivation object is assigned to the
global environment with the name \code{derivation_name}.
}
\description{
Load a derivation's output into global environment
}
