% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_derivation.R
\name{drv_quarto}
\alias{drv_quarto}
\title{Render a Quarto document as a Nix derivation}
\usage{
drv_quarto(name, qmd_file)
}
\arguments{
\item{name}{Symbol, derivation name.}

\item{qmd_file}{Character, path to .qmd file.}
}
\value{
List with \code{name} (string) and \code{snippet} (Nix code).
}
\description{
Render a Quarto document as a Nix derivation
}
\details{
Detects \code{drv_read("ref")} in the .qmd file and replaces with derivation output paths.
}
\examples{
drv_quarto(report, "doc.qmd")

}
