% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deriv_read_load.R
\name{drv_read}
\alias{drv_read}
\title{Read output of a derivation}
\usage{
drv_read(derivation_name, result_path = "_rixpress")
}
\arguments{
\item{derivation_name}{Character, the name of the derivation.}
}
\value{
The derivation's output.
}
\description{
Read output of a derivation
}
