% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxp_make.R
\name{export_nix_archive}
\alias{export_nix_archive}
\title{Export Nix store paths to an archive}
\usage{
export_nix_archive(
  store_paths = "_rixpress",
  archive_file = "_rixpress/pipeline-outputs.nar"
)
}
\arguments{
\item{store_paths}{Character, defaults to "_rixpress", the folder holding the
result symlinks.}

\item{archive_file}{Character, path to the archive, defaults to
"_rixpress/pipeline-outputs.nar"}
}
\description{
Creates a single archive file containing the specified Nix store paths
and their dependencies.
This archive can be transferred to another machine and imported into
its Nix store.
}
