% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxp_write_dag.R
\name{extract_markdown_dependencies}
\alias{extract_markdown_dependencies}
\title{Extract dependencies from markdown (Quarto/RMarkdown) derivations}
\usage{
extract_markdown_dependencies(deriv, type, name, all_derivs_names)
}
\arguments{
\item{deriv}{The derivation object}

\item{type}{The type of derivation ("rxp_qmd" or "rxp_rmd")}

\item{name}{The name of the derivation}

\item{all_derivs_names}{All derivation names}
}
\value{
A character vector of dependency names
}
\description{
Extract dependencies from markdown (Quarto/RMarkdown) derivations
}
\keyword{internal}
