% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxp_write_dag.R
\name{extract_r_dependencies}
\alias{extract_r_dependencies}
\title{Extract dependencies from R script derivations}
\usage{
extract_r_dependencies(deriv, name, all_derivs_names)
}
\arguments{
\item{deriv}{The derivation object}

\item{name}{The name of the derivation}

\item{all_derivs_names}{All derivation names}
}
\value{
A character vector of dependency names
}
\description{
Extract dependencies from R script derivations
}
\keyword{internal}
