% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dag.R
\name{generate_dag}
\alias{generate_dag}
\title{Generate a DAG from a list of derivations}
\usage{
generate_dag(deriv_list, output_file = "_rixpress/dag.json")
}
\arguments{
\item{deriv_list}{A list of derivations, each with a \code{name} and \code{snippet}, output of derive_r().}

\item{output_file}{Path to the output JSON file. Defaults to "_rixpress/dag.json".}
}
\value{
Writes a JSON file representing the DAG.
}
\description{
Creates a JSON representation of a directed acyclic graph (DAG) based on dependencies
between derivations.
}
