% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_pkgs.R
\name{generate_libraries_script}
\alias{generate_libraries_script}
\title{Generate a script with import statements and optional additional content}
\usage{
generate_libraries_script(
  packages,
  additional_files,
  outfile,
  import_formatter,
  additional_file_pattern
)
}
\arguments{
\item{packages}{List of package names}

\item{additional_files}{Character vector of additional files to include}

\item{outfile}{Path to the output file}

\item{import_formatter}{Function to format import statements}

\item{additional_file_pattern}{Regex pattern to identify additional files}
}
\value{
Writes a script to the specified outfile and returns the path
}
\description{
Generate a script with import statements and optional additional content
}
