% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivs.R
\name{generate_substitution_commands}
\alias{generate_substitution_commands}
\title{Generate substitution commands for document references}
\usage{
generate_substitution_commands(all_matches, doc_file)
}
\arguments{
\item{all_matches}{Data frame with match information}

\item{doc_file}{Document file name}
}
\value{
Character vector of substitution commands
}
\description{
Generate substitution commands for document references
}
\keyword{internal}
