% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rixpress.R
\name{get_need_jl}
\alias{get_need_jl}
\title{Check if Julia is needed based on derivation types}
\usage{
get_need_jl(types)
}
\arguments{
\item{types}{Character vector of derivation types}
}
\value{
Logical indicating if Julia is needed
}
\description{
Check if Julia is needed based on derivation types
}
