% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rixpress.R
\name{get_need_py}
\alias{get_need_py}
\title{Check if Python is needed based on derivation types}
\usage{
get_need_py(types)
}
\arguments{
\item{types}{Character vector of derivation types}
}
\value{
Logical indicating if Python is needed
}
\description{
Check if Python is needed based on derivation types
}
