% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rixpress.R
\name{get_need_r}
\alias{get_need_r}
\title{Check if R is needed based on derivation types}
\usage{
get_need_r(types)
}
\arguments{
\item{types}{Character vector of derivation types}
}
\value{
Logical indicating if R is needed
}
\description{
Check if R is needed based on derivation types
}
