% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dag.R
\name{get_nodes_edges}
\alias{get_nodes_edges}
\title{Prepare data for plotting the DAG of the pipeline.}
\usage{
get_nodes_edges(path_dag = "_rixpress/dag.json")
}
\arguments{
\item{path_dag}{Path to the DAG JSON file}
}
\value{
A list of two datasets, \code{nodes} and \code{edges}.

List with nodes and edges data frames
}
\description{
Prepare data for plotting the DAG of the pipeline.
}
\examples{
\dontrun{
  get_nodes_edges()
}
}
