% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_pkgs.R
\name{import_formatter_py}
\alias{import_formatter_py}
\title{Format Python import statements}
\usage{
import_formatter_py(package)
}
\arguments{
\item{package}{Character, name of the package}
}
\value{
Character string with import statement
}
\description{
Format Python import statements
}
