% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_pkgs.R
\name{import_formatter_r}
\alias{import_formatter_r}
\title{Format R import statements}
\usage{
import_formatter_r(package)
}
\arguments{
\item{package}{Character, name of the package}
}
\value{
Character string with import statement
}
\description{
Format R import statements
}
