% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxp_make.R
\name{import_nix_archive}
\alias{import_nix_archive}
\title{Import Nix store paths from an archive}
\usage{
import_nix_archive(archive_file)
}
\arguments{
\item{archive_file}{The file name of the archive to import
(e.g., "my_archive.nar").}
}
\description{
Imports the store paths contained in an archive file into the local Nix store.
Useful for transferring built outputs between machines.
}
