% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxp_io.R
\name{is_remote_url}
\alias{is_remote_url}
\title{Check if path is a remote URL}
\usage{
is_remote_url(path)
}
\arguments{
\item{path}{Character path}
}
\value{
Logical
}
\description{
Check if path is a remote URL
}
\keyword{internal}
