% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rixpress.R
\name{parse_nix_envs}
\alias{parse_nix_envs}
\title{parse_nix_envs Parses the nix_env element of a deriv object}
\usage{
parse_nix_envs(derivs)
}
\arguments{
\item{derivs}{List of derivation objects}
}
\value{
List of unique Nix environments
}
\description{
parse_nix_envs Parses the nix_env element of a deriv object
}
