% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivs.R
\name{print.rxp_derivation}
\alias{print.rxp_derivation}
\title{Print Method for Derivation Objects}
\usage{
\method{print}{rxp_derivation}(x, ...)
}
\arguments{
\item{x}{An object of class "rxp_derivation"}

\item{...}{Additional arguments passed to print methods}
}
\value{
Nothing, prints a summary of the derivation object to the console.
}
\description{
Print Method for Derivation Objects
}
\examples{
\dontrun{
# d0 is a previously defined derivation
  print(d0)
}
}
\seealso{
Other utilities: 
\code{\link{rxp_check_chronicles}()},
\code{\link{rxp_copy}()},
\code{\link{rxp_gc}()},
\code{\link{rxp_init}()},
\code{\link{rxp_inspect}()},
\code{\link{rxp_list_logs}()},
\code{\link{rxp_load}()},
\code{\link{rxp_read}()},
\code{\link{rxp_trace}()}
}
\concept{utilities}
