% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxp_pipeline.R
\name{print.rxp_pipeline}
\alias{print.rxp_pipeline}
\title{Print Method for rxp_pipeline Objects}
\usage{
\method{print}{rxp_pipeline}(x, ...)
}
\arguments{
\item{x}{An object of class "rxp_pipeline"}

\item{...}{Additional arguments passed to print methods}
}
\value{
Nothing, prints a summary of the pipeline object to the console.
}
\description{
Print Method for rxp_pipeline Objects
}
