% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxp_io.R
\name{process_read_function}
\alias{process_read_function}
\title{Process read function for different languages}
\usage{
process_read_function(read_function, lang, parent_env = parent.frame())
}
\arguments{
\item{read_function}{Function or character}

\item{lang}{Language string}

\item{parent_env}{Environment from calling function for proper substitution}
}
\value{
Character string
}
\description{
Process read function for different languages
}
\keyword{internal}
