% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxp_write_dag.R
\name{propagate_noop_builds}
\alias{propagate_noop_builds}
\title{Propagate no-op builds to dependent derivations}
\usage{
propagate_noop_builds(dag)
}
\arguments{
\item{dag}{A list of derivation nodes with dependencies}
}
\value{
Updated DAG with propagated no-op flags
}
\description{
Propagate no-op builds to dependent derivations
}
\keyword{internal}
