% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxp_io.R
\name{rxp_common_setup}
\alias{rxp_common_setup}
\title{Generate the Nix derivation snippet for Python-R object transfer.}
\usage{
rxp_common_setup(out_name, expr_str, nix_env, direction)
}
\arguments{
\item{out_name}{Character, name of the derivation.}

\item{expr_str}{Character, name of the object being transferred.}

\item{nix_env}{Character, path to the Nix environment file.}

\item{direction}{Character, either "py2r" (Python to R) or "r2py" (R to
Python).}
}
\value{
A list with elements: \code{name}, \code{snippet}, \code{type}, \code{additional_files},
\code{nix_env}.
}
\description{
This function constructs the \code{build_phase} and Nix derivation snippet
based on the given parameters.
}
\keyword{internal}
