% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxp_copy.R
\name{rxp_copy}
\alias{rxp_copy}
\title{Copy derivations from Nix store to current working directory}
\usage{
rxp_copy(derivation_name = NULL)
}
\arguments{
\item{derivation_name}{The name of the derivation to copy. If empty, then all
the derivations are copied.}
}
\value{
Nothing, the contents of the Nix store are copied to the current working directory.
}
\description{
When Nix builds a derivation, its output is save the "Nix store"
located under \verb{/nix/store/}. Even though you can import the derivations
into the current R session using \code{rxp_read()} or \code{rxp_load()}, it can be
useful to copy the outputs to the current working directory. This is especially
useful for Quarto documents, where there can be more than one input, as is
the case for \code{html} output.
}
