% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxp_init.R
\name{rxp_init}
\alias{rxp_init}
\title{Initialize pipeline}
\usage{
rxp_init(project_path, overwrite = FALSE)
}
\arguments{
\item{project_path}{Character string specifying the project's path.}

\item{overwrite}{Logical, defaults to FALSE, indicating whether to overwrite
existing scripts.}
}
\description{
Generates \code{gen-env.R} and \code{gen-pipeline.R} scripts in the specified project
directory.
}
\details{
Creates:
\itemize{
\item \code{gen-env.R}: Script to define an execution environment with \code{{rix}}.
\item \code{gen-pipeline.R}: Defines a data pipeline with \code{{rixpress}}.
}
}
\examples{
# Default usage
rxp_init()

# Custom path with overwrite
rxp_init(project_path = "my_project", overwrite = TRUE)

}
