% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxp_inspect.R
\name{rxp_inspect}
\alias{rxp_inspect}
\title{Inspect the build result of a pipeline.}
\usage{
rxp_inspect()
}
\value{
A data frame with derivation names, if their build was successful,
their paths in the /nix/store, and their build outputs.
}
\description{
Returns a data frame with four columns:
- derivation: the name of the derivation
- build_success: whether the build was successful or not
- path: the path of this derivation in the Nix store
- output: the output, if this derivation was built successfully.
Empty outputs mean that this derivation was not built
successfully. Several outputs for a single derivation
are possible.
}
