% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxp_inspect.R
\name{rxp_list_logs}
\alias{rxp_list_logs}
\title{List all available build logs}
\usage{
rxp_list_logs(project_path = ".")
}
\arguments{
\item{project_path}{Character, defaults to ".".
Path to the root directory of the project.}
}
\value{
A data frame with log filenames, modification times, and file sizes.
}
\description{
Returns a data frame with information about all build logs
in the project's _rixpress directory.
}
