% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxp_make.R
\name{rxp_make}
\alias{rxp_make}
\title{Build pipeline using Nix}
\usage{
rxp_make(verbose = FALSE)
}
\arguments{
\item{verbose}{Logical, defaults to FALSE. Set to TRUE to see nix's
standard output, can be useful to check what is happening if the
build process takes long.}
}
\value{
A character vector of paths to the built outputs.
}
\description{
Runs \code{nix-build} with a quiet flag, outputting to \verb{_rixpress/result}.
}
