% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivs.R
\name{rxp_py2r}
\alias{rxp_py2r}
\title{Transfer Python object into an R session.}
\usage{
rxp_py2r(name, expr, nix_env = "default.nix")
}
\arguments{
\item{name}{Symbol, name of the derivation.}

\item{expr}{Symbol, Python object to be loaded into R.}

\item{nix_env}{Character, path to the Nix environment file, default is
"default.nix".}
}
\value{
An object of class derivation which inherits from lists.
}
\description{
Transfer Python object into an R session.
}
\details{
\code{rxp_py2r(my_obj, my_python_object)} loads a serialized Python
object and saves it as an RDS file using \code{reticulate::py_load_object()}.
}
\examples{
\dontrun{
rxp_py2r(my_obj, my_python_object)
}
}
\seealso{
Other derivations: 
\code{\link{rxp_file_common}()},
\code{\link{rxp_py_file}()},
\code{\link{rxp_quarto}()},
\code{\link{rxp_r}()},
\code{\link{rxp_r2py}()},
\code{\link{rxp_rmd}()}
}
\concept{derivations}
