% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivs.R
\name{rxp_r}
\alias{rxp_r}
\title{rxp_r Creates a Nix expression running an R function}
\usage{
rxp_r(name, expr, additional_files = "", nix_env = "default.nix")
}
\arguments{
\item{name}{Symbol, name of the derivation.}

\item{expr}{R code to generate the expression.}

\item{additional_files}{Character vector, additional files to include. These
are the files that contain custom functions required for this derivation.}

\item{nix_env}{Character, path to the Nix environment file, default is "default.nix".}

\item{output}{Character, a variable name to save the
output of the function expression}
}
\value{
A list with elements: \code{name}, the \code{name} of the derivation,
\code{snippet}, the Nix boilerplate code, \code{type}, and \code{nix_env}.
}
\description{
rxp_r Creates a Nix expression running an R function
}
\details{
At a basic level, \code{rxp_r(mtcars_am, filter(mtcars, am == 1))}
is equivalent to \code{mtcars <- filter(mtcars, am == 1)}. \code{rxp_r()} generates
the required Nix boilerplate to output a so-called "derivation" in Nix
jargon. A Nix derivation is a recipe that defines how to create an output
(in this case \code{mtcars_am}) including its dependencies, build steps,
and output paths.
}
\examples{
rxp_r(mtcars_am, filter(mtcars, am == 1))
}
