% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxp_io.R
\name{rxp_r2py}
\alias{rxp_r2py}
\title{Transfer R Object into a Python Session}
\usage{
rxp_r2py(name, expr, nix_env = "default.nix", env_var = NULL)
}
\arguments{
\item{name}{Symbol, name of the derivation.}

\item{expr}{Symbol, R object to be saved into a Python pickle.}

\item{nix_env}{Character, path to the Nix environment file, default is
"default.nix".}

\item{env_var}{Named list of environment variables to set. RETICULATE_AUTOCONFIGURE=0
is set by default to prevent reticulate from modifying PYTHONPATH in Nix builds.}
}
\value{
An object of class \code{rxp_derivation}.
}
\description{
Transfer R Object into a Python Session
}
\details{
\code{rxp_r2py(my_obj, my_r_object)} saves an R object to a Python pickle
using \code{reticulate::py_save_object()}.
}
\seealso{
Other interop functions: 
\code{\link{rxp_py2r}()}
}
\concept{interop functions}
