% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivs.R
\name{rxp_r_file}
\alias{rxp_r_file}
\title{rxp_r_file Creates a Nix expression that reads in a file using R.}
\usage{
rxp_r_file(name, path, read_function, nix_env = "default.nix")
}
\arguments{
\item{name}{Symbol, the name of the derivation.}

\item{path}{Character, the file path to include, can be a URL.}

\item{read_function}{Function, an R function to read the data, taking one argument (the path).}

\item{nix_env}{Character, path to the Nix environment file, default is "default.nix".}
}
\value{
A list with \code{name}, \code{snippet}, \code{type}, and \code{nix_env}.
}
\description{
rxp_r_file Creates a Nix expression that reads in a file using R.
}
\details{
See original documentation for usage notes.
}
