% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deriv_read_load.R
\name{rxp_read}
\alias{rxp_read}
\title{Read output of a derivation}
\usage{
rxp_read(derivation_name, result_path = "_rixpress")
}
\arguments{
\item{derivation_name}{Character, the name of the derivation.}

\item{result_path}{The folder containing the pipelines' outputs.}
}
\value{
The derivation's output.
}
\description{
Read output of a derivation
}
