% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivs.R
\name{sanitize_nix_base}
\alias{sanitize_nix_base}
\title{Sanitize Nix environment name to valid identifier}
\usage{
sanitize_nix_base(nix_env)
}
\arguments{
\item{nix_env}{Nix environment file path}
}
\value{
Sanitized base name
}
\description{
Sanitize Nix environment name to valid identifier
}
\keyword{internal}
