% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_pkgs.R
\name{transform_r}
\alias{transform_r}
\title{Transform R package names from Nix format to CRAN format}
\usage{
transform_r(packages)
}
\arguments{
\item{packages}{Character vector of package names}
}
\value{
Character vector of transformed package names
}
\description{
Transform R package names from Nix format to CRAN format
}
