% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dag.R
\name{unnest_all_columns}
\alias{unnest_all_columns}
\title{Unnest all columns in a data frame}
\usage{
unnest_all_columns(df)
}
\arguments{
\item{df}{Data frame to unnest}
}
\value{
Unnested data frame
}
\description{
Unnest all columns in a data frame
}
