% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxp_io.R
\name{build_language_commands}
\alias{build_language_commands}
\title{Build language-specific execution commands}
\usage{
build_language_commands(
  lang,
  read_func,
  user_code,
  out_name,
  rel_path,
  serialize_str = NULL
)
}
\arguments{
\item{lang}{Language string}

\item{read_func}{String representing the function to call}

\item{user_code}{Source/import statements for user functions}

\item{out_name}{Name of the output object}

\item{rel_path}{Relative path to input file}

\item{serialize_str}{Language-specific serializer identifier or NULL for default}
}
\value{
Character vector of commands
}
\description{
Build language-specific execution commands
}
\keyword{internal}
