% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxp_make.R
\name{export_nix_archive}
\alias{export_nix_archive}
\title{Export Nix store paths to an archive}
\usage{
export_nix_archive(archive_file = "_rixpress/pipeline_outputs.nar")
}
\arguments{
\item{archive_file}{Character, path to the archive, defaults to
"_rixpress/pipeline-outputs.nar"}
}
\description{
Creates a single archive file containing the specified Nix store paths
and their dependencies.
This archive can be transferred to another machine and imported into
its Nix store.
}
