% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivs.R
\name{print.derivation}
\alias{print.derivation}
\title{Print method for derivation objects}
\usage{
\method{print}{derivation}(x, ...)
}
\arguments{
\item{x}{An object of class "derivation"}

\item{...}{Additional arguments passed to print methods}
}
\value{
Nothing, prints a summary of the derivation object to the console.
}
\description{
Print method for derivation objects
}
\examples{
\dontrun{
# d0 is a previously defined derivation
  print(d0)
}
}
