% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivs.R
\name{rxp_file}
\alias{rxp_file}
\title{rxp_file Creates a Nix expression that reads in a file.}
\usage{
rxp_file(name, path, read_function, nix_env = "default.nix")
}
\arguments{
\item{name}{Symbol, the name of the derivation.}

\item{path}{Character, the file path to include, can be an url.}

\item{read_function}{Function to read in the data, must be a
function of only one argument, the path. If you wish to pass
several arguments to this function, make it an anonymous function.}

\item{nix_env}{Character, path to the Nix environment file, default is "default.nix".}
}
\value{
A list with elements: \code{name}, the derivation name,
\code{snippet}, the generated Nix code, \code{type}, and \code{nix_env}.
}
\description{
rxp_file Creates a Nix expression that reads in a file.
}
\details{
The function must only take one single argument, the
path to the file to read. Because of this limitation, if
you need to pass more than one argument to the function,
make it an anonymous function, for example:
\verb{d0 <- rxp_file(mtcars, 'mtcars.csv', \\(x) (read.csv(file = x, sep = "|")))}
}
\examples{
\dontrun{
rxp_file(data, "./data.csv", read.csv)
}
}
