% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxp_ga.R
\name{rxp_ga}
\alias{rxp_ga}
\title{Run a pipeline on GitHub Actions}
\usage{
rxp_ga()
}
\value{
Nothing, copies file to a directory.
}
\description{
Run a pipeline on GitHub Actions
}
\details{
This function puts a \code{.yaml} file inside the \verb{.github/workflows/}
folder on the root of your project. This workflow file expects both
scripts generated by \code{rxp_init()}, \code{gen-env.R} and \code{gen-pipeline.R} to be
present. If that's not the case, edit the \code{.yaml} file accordingly. Build
artifacts are archived and restored automatically between runs. Make sure
to give read and write permissions to the GitHub Actions bot.
}
\examples{
\dontrun{
  rxp_ga()
}
}
\seealso{
Other ci utilities: 
\code{\link{rxp_dag_for_ci}()},
\code{\link{rxp_write_dag}()}
}
\concept{ci utilities}
