% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dag.R
\name{rxp_ggdag}
\alias{rxp_ggdag}
\title{Create a Directed Acyclic Graph (DAG) representing the pipeline
using \code{{ggplot2}}}
\usage{
rxp_ggdag(nodes_and_edges = get_nodes_edges())
}
\arguments{
\item{nodes_and_edges}{List, output of \code{get_nodes_edges()}.}
}
\value{
A \code{{ggplot2}} object.
}
\description{
Uses \code{{ggdag}} to generate the plot. \code{{ggdag}} is a soft
dependency of \code{{rixpress}} so you need to install it to use this
function.
}
\examples{
\dontrun{
  rxp_ggdag()
}
}
\seealso{
Other visualisation functions: 
\code{\link{rxp_visnetwork}()}
}
\concept{visualisation functions}
