% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivs.R
\name{rxp_py2r}
\alias{rxp_py2r}
\title{Transfer Python object into an R session.}
\usage{
rxp_py2r(name, expr, additional_files = "", nix_env = "default.nix")
}
\arguments{
\item{name}{Symbol, name of the derivation.}

\item{expr}{Python object to be loaded into R.}

\item{additional_files}{Character vector, additional files to include. These
are the files that contain custom functions required for this derivation.}

\item{nix_env}{Character, path to the Nix environment file, default is "default.nix".}
}
\value{
A list with elements: \code{name}, the \code{name} of the derivation,
\code{snippet}, the Nix boilerplate code, \code{type}, and \code{nix_env}.
}
\description{
Transfer Python object into an R session.
}
\details{
\code{rxp_py2r(my_obj, "my_python_object")} loads a serialized
Python object and saves it as an RDS file using \code{reticulate::py_load_object()}.
}
\examples{
rxp_py2r(my_obj, "my_python_object")
}
