% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxp_io.R
\name{rxp_py2r}
\alias{rxp_py2r}
\title{Transfer Python object into an R session.}
\usage{
rxp_py2r(name, expr, nix_env = "default.nix")
}
\arguments{
\item{name}{Symbol, name of the derivation.}

\item{expr}{Symbol, Python object to be loaded into R.}

\item{nix_env}{Character, path to the Nix environment file, default is
"default.nix".}
}
\value{
An object of class \code{rxp_derivation}.
}
\description{
Transfer Python object into an R session.
}
\details{
\code{rxp_py2r(my_obj, my_python_object)} loads a serialized Python
object and saves it as an RDS file using \code{reticulate::py_load_object()}.
}
\seealso{
Other interop functions: 
\code{\link{rxp_r2py}()}
}
\concept{interop functions}
