% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxp_io.R
\name{rxp_py_file}
\alias{rxp_py_file}
\title{Creates a Nix expression that reads in a file (or folder of data) using Python.}
\usage{
rxp_py_file(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=rxp_file]{rxp_file}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
An object of class \code{rxp_derivation}.
}
\description{
Creates a Nix expression that reads in a file (or folder of data) using Python.
}
\seealso{
Other derivations: 
\code{\link{rxp_jl}()},
\code{\link{rxp_jl_file}()},
\code{\link{rxp_py}()},
\code{\link{rxp_qmd}()},
\code{\link{rxp_r}()},
\code{\link{rxp_r_file}()},
\code{\link{rxp_rmd}()}
}
\concept{derivations}
