% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxp_io.R
\name{rxp_r2py}
\alias{rxp_r2py}
\title{Transfer R object into a Python session.}
\usage{
rxp_r2py(name, expr, nix_env = "default.nix")
}
\arguments{
\item{name}{Symbol, name of the derivation.}

\item{expr}{Symbol, R object to be saved into a Python pickle.}

\item{nix_env}{Character, path to the Nix environment file, default is
"default.nix".}
}
\value{
An object of class \code{rxp_derivation}.
}
\description{
Transfer R object into a Python session.
}
\details{
\code{rxp_r2py(my_obj, my_r_object)} saves an R object to a Python pickle
using \code{reticulate::py_save_object()}.
}
\seealso{
Other interop functions: 
\code{\link{rxp_py2r}()}
}
\concept{interop functions}
