% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxp_io.R
\name{rxp_r_file}
\alias{rxp_r_file}
\title{Creates a Nix expression that reads in a file (or folder of data) using R.}
\usage{
rxp_r_file(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=rxp_file]{rxp_file}}
  \describe{
    \item{\code{}}{}
  }}

\item{name}{Symbol, the name of the derivation.}

\item{path}{Character, file or folder path to include.}

\item{read_function}{Function, R function to read the data.}

\item{user_functions}{Character vector of script paths to include.}

\item{nix_env}{Character, path to the Nix environment file.}

\item{env_var}{Named list of environment variables.}
}
\value{
An object of class \code{rxp_derivation}.
}
\description{
Creates a Nix expression that reads in a file (or folder of data) using R.
}
\seealso{
Other derivations: 
\code{\link{rxp_jl}()},
\code{\link{rxp_jl_file}()},
\code{\link{rxp_py}()},
\code{\link{rxp_py_file}()},
\code{\link{rxp_qmd}()},
\code{\link{rxp_r}()},
\code{\link{rxp_rmd}()}
}
\concept{derivations}
