% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxp_read_load.R
\name{rxp_read}
\alias{rxp_read}
\title{Read output of a derivation}
\usage{
rxp_read(derivation_name)
}
\arguments{
\item{derivation_name}{Character, the name of the derivation.}
}
\value{
The derivation's output.
}
\description{
Reads the output of derivations in the current session,
returns a path if reading directly is not possible.
}
\details{
When \code{derivation_name} points to a single R object,
it gets read in the current session using \code{readRDS()}.
If it’s a Python object and {reticulate} is available,
\code{reticulate::py_load_object()} is used. In case
the derivation is pointing to several outputs (which can
happen when building a Quarto document for example) or
neither \code{readRDS()} nor \code{reticulate::py_load_object()}
successfully read the object, the path to the object is
returned instead.
}
\examples{
\dontrun{
  mtcars <- rxp_read("mtcars")
}
}
